Refactor main.js.
Changes include:
- Remove items in main.js and create their own modules in components folder
- Delete main.js
- Add header, footer, and cookies to each individual js file for each page.
- Update script tags on all pages.
