+++ index.html
@@ -81,7 +81,6 @@
     <footer
       class="background-black curved-dark-before container footer"
     ></footer>
-    <script src="js/main.js"></script>
     <script type="module" src="js/index.js"></script>
   </body>
 </html>
