+++ js/category-page.js
@@ -1,0 +1,7 @@
+import { buildNavigation } from "./component/header.js";
+import { buildFooter } from "./component/footer.js";
+import { setCookies } from "./component/cookieConsent.js";
+
+buildNavigation();
+buildFooter();
+setCookies();
