+++ js/locals-guide-landing.js
@@ -1,3 +1,7 @@
+import { buildNavigation } from "./component/header.js";
+import { buildFooter } from "./component/footer.js";
+import { setCookies } from "./component/cookieConsent.js";
+
 const buttons = document.querySelectorAll(".lg-button-container button");
 
 buttons.forEach((button) => {
@@ -5,3 +9,7 @@ buttons.forEach((button) => {
     window.location.href = `/locals-guide/${button.innerHTML.toLocaleLowerCase()}.html`;
   });
 });
+
+buildNavigation();
+buildFooter();
+setCookies();
