+++ locals-guide/drink.html
@@ -20,6 +20,6 @@
     <footer
       class="background-black curved-dark-before container footer"
     ></footer>
-    <script src="/js/main.js"></script>
+    <script type="module" src="/js/category-page.js"></script>
   </body>
 </html>
