+++ locals-guide/index.html
@@ -27,7 +27,6 @@
     <footer
       class="background-black curved-dark-before container footer"
     ></footer>
-    <script src="/js/main.js"></script>
-    <script src="/js/locals-guide-landing.js"></script>
+    <script type="module" src="/js/locals-guide-landing.js"></script>
   </body>
 </html>
